(maint) Add Docker on Windows support
Previously the docker compose would fail due to permission issues with
Postgres and shared volumes.  This commit adds a docker compose override file
and instructions on how to use PupperWare on Windows.

This commit also adds a PowerShell equivalent to the bash file so that puppet
commands can be sent to the stack on Windows.