(maint) Update building instructions for Windows 10
Previously the installation instructions did not stress that the LCOW building
process reqquired a working Docker installation.  Due to LCOW being
non-functional this meant the user could never update the LCOW VM.  This commit
adds the a warning text and recommends installing Docker For Windows (Docker
Desktop) as well to facilitate the linuxkit building process.