Experiment with a special "migration" container
The goal is to figure out a pattern where we can run some code prior to
the production services starting up, which would allow us to perform
"migrations" from earlier versions of the compose file to the current
version. Essentially I'm after a "pre-up" phase that runs before the
normal "docker-compose up" phase.

This would be handy when we do things like move the location of
volume-mounted directories. The special migration container would detect
the old locations and move them into the new locations.

I'm not sure that we can reliably use "depends_on" and docker-compose
for this, because there's no guarantee that the migration container is
done running before the other containers start up. It currently works
because these are simple mv commands which happen to be fast enough.

The better solution is to bake in special "wait-for" support into each
production service/image.