Build an HAProxy container which we will use as the laod balancer
We will have consul-template regenerate and reload the HAProxy config every
time a compiler is added or removed.

The container waits for the proxy cert (master's public and private cert)
to show up before starting. The cert will be generated by the master.

The HAProxy config sends all CA traffic to 'master', and balances all other
traffic between 'master' and additional compilers; since the CA is
low-traffic, we use it for catalog compilation etc., too.