Add HAProxy in front of the master and addl compilers
We add our lb container, together with an rsyslog container, since HAProxy
will not log to stdout, and logging via rsyslog is the recommended setup
for containers.

We call the lb 'puppet' so that all traffic, including internal traffic
goes through this by default. The puppetserver with the CA is called
'master' whereas additional compilers are called 'compiler'.

Since we terminate SSL in HAProxy, we reconfigure all puppetservers to
accept HTTP traffic, and use HTTP headers to identify clients. (That's very
dangerous without the lb in front)

The docker-entrypoint.sh for puppetserver also explicitly generates a CA
and master cert if none exists, and brings it into a form that HAProxy
likes in /etc/puppetlabs/puppet/ssl/proxy_cert.pem

On compilers, we turn off the CA completely, and make sure we register and
deregister with consul when the container is started/stopped.