(MAINT) SSL certs without Ruby, Puppet
The intention is to remove the need for Ruby and Puppet in order to get
SSL certs.

This simplifies the requirements for "dockerizing" an application since
openssl and curl are available in many places that ruby and the
puppet-agent aren't.

This script should be downloaded via "ADD <url>" in any Dockerfile,
and then executed in the application entrypoint script.