(MAINT) Fail fast if CSR already exists
Previously the script would print out that a CSR already exists and then
continue on attempting to fetch the signed CSR, eventually hitting the
WAIT_FOR_CERT timeout and then failing.

This commit adds a check after the CSR submission and fails the script
at that point if there's already a CSR on the CA.

We could even move this check further up the script and before anything
is generated on the client by attempting to pre-fetch a CSR for this
host and failing if one is returned.