(MAINT) Support DNS alt names in ssl.sh
This relies on the the `-addext` flag on openssl, which is not available
on some versions of "openssl", such as the default one on Mac OS (which
may actually be LibreSSL).

If we need to, we can implement this a different way, by instead writing
extensions to a .conf file and then specifying that on the CLI via
`-config .conf`. That should work on more versions of "openssl".

Note that the Puppet CA does not allow DNS alt names by default and will
not sign the certificate request.