(MAINT) Enable admin login
This is a convenience and ease-of-use change. We'll want to revisit this
before we ship.

Currently the PE stack is not accessible when it comes up, you need to
run the `make admin-login` SQL command that we have copied around, or
call the special ruby helper method to run the SQL command.

This moves that SQL command into the container itself, so we don't have
to perform the manual step anymore.