(MAINT) Set PGDATA to avoid Windows problem
By default postgres tries creating the PGDATA directory under its own
home directory, and it relies on Unix-only permission bits to give
itself permission to do that. This fails on Windows, as expected.

The solution is to either explicitly chmod/chown the home
directory (what we were doing before, with volume mapping), or just set
the directory to be somewhere else where it already has permission.