(maint) Add two helper methods for pe-bolt-server
curl_console_task: This will hit the console endpoint that will generate
a job that will run a task over SSH. The task needs to be in the puppetserver's
module path in order for this to work. pe-bolt-server's docker directory has
an example of this.
curl_job_number: curls the console for a job until the job is no longer running.