(MAINT) Don't redirect in exec command
Using pipes, redirects in a `docker-compose exec <COMMAND>` invocation
is problematic as the pipe/redirect will not be considered part of the
COMMAND and thus won't run in the container, but instead on the host
system.

In this case, just remove the redirect and let the `No such file or
directory` message appear in the test output, since this change was
simply for quieting down errors/warnings in the test output.