(MAINT) Strip gem down to what we use
Remove dependencies and local convenience files so the gem is smaller
and faster to download (via bundler).

This is a step towards sharing a common Makefile, assuming we distribute
the common Makefile via our gem.

Since we don't have any tests for this gem/repo, and we're not
publishing the gem, we don't need the bundler, rake, or rspec
dependencies, and json is already available in core Ruby.