(REPLATS-150) Set service uid/gid labels for cert dirs
The SpecHelpers import preset certs to speed up start up time. To do
this they use a trick of composing the containers with --no-start first
so their entrypoints don't run but volumes are created, then running a
base alpine container with the data dir volume mounted and with the
preset certs mounted, and then copying over the certs and setting
permissions.

All of that is fine, but the act of mounting the datadir in alpine under
/opt (which is owned by root in that container) ends up setting the
host volume path's uid to root as well, which then later interferes with
the container's service user interacting with the datadir when the
actual service container is brought up with the volume mounted.

With this patch, the mount point is instead chowned with the labeled
uid/gid, ensuring the volume root continues to be owned by the service
user. The service ids match those set in the individual service container
repo Dockerfiles.