(PE-32949) Explicitly run test cert gen script via `/bin/sh`
We have a script that is used the in pe-puppet-server-extensions
container tests to generate certs for specific certs names. It is run in
an ephemeral container, based on the `alpine/openssl` image. When the
script was updated to run under bash, the helper to run it in the temp
container broke, because bash is not available in alpine. This commit
updates that docker invocation to explicitly run the script in `sh`.